backend/pyproject.toml
[project]
name = "dewey"
version = "0.1.0"
description = "AI-Powered Communication Processing Platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "wfdeller"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",

    # Database
    "sqlmodel>=0.0.14",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",

    # Redis
    "redis>=5.0.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "argon2-cffi>=23.1.0",

    # AI Providers
    "anthropic>=0.18.0",
    "openai>=1.10.0",
    "httpx>=0.26.0",

    # Email Processing
    "aiosmtplib>=3.0.0",
    "aioimaplib>=1.0.1",
    "email-validator>=2.1.0",

    # Microsoft Graph API
    "msal>=1.26.0",
    "msgraph-sdk>=1.2.0",

    # Data Processing
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "testcontainers[postgres,redis]>=3.7.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",

    # Load Testing
    "locust>=2.22.0",

    # Code Quality
    "ruff>=0.2.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",

    # Type Stubs
    "types-redis>=4.6.0",
    "types-passlib>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG001"]
"alembic/*" = ["ARG001"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "aioimaplib.*",
    "msal.*",
    "msgraph.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
